pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic==1.16.5",
    "blinker==1.9.0",
    "click==8.3.0",
    "colorama==0.4.6",
    "distlib==0.4.0",
    "filelock==3.19.1",
    "flask==3.1.2",
    "flask-cors>=6.0.1",
    "flask-marshmallow==1.3.0",
    "flask-migrate==4.1.0",
    "flask-sqlalchemy==3.1.1",
    "greenlet==3.2.4",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "mako==1.3.10",
    "markupsafe==3.0.2",
    "marshmallow==4.0.1",
    "marshmallow-sqlalchemy==1.4.2",
    "platformdirs==4.4.0",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "setuptools==80.9.0",
    "sqlalchemy==2.0.43",
    "typing-extensions==4.15.0",
    "virtualenv==20.34.0",
    "werkzeug==3.1.3",
    "wheel==0.45.1",
]
